frontend Settings: Refactoring to follow component conventions
