--- frontend/src/components/App/settings/config.ts
@@ -1,1 +1,0 @@
-export const DefaultRowsPerPageOptions = [15, 25, 50];
