+++ frontend/src/components/ingress/Details.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { useParams } from 'react-router-dom';
 import Ingress from '../../lib/k8s/ingress';
-import { useSettings } from '../App/settings/hook';
+import { useSettings } from '../App/Settings/hook';
 import { DetailsGrid } from '../common/Resource';
 import { SectionBox } from '../common/SectionBox';
 import SimpleTable from '../common/SimpleTable';
